+++ anywise/demo/app.py
@@ -1,10 +1,27 @@
 import typing as ty
 
 import uvicorn
-from fastapi import FastAPI
+from fastapi import APIRouter, FastAPI
+from sqlalchemy.ext.asyncio import AsyncEngine
 
-from ..anywise import Anywise
-from .todo import todo_router
+from .. import Anywise
+from ..integration.fastapi import FastWise
+from .db import create_tables
+from .model import CreateTodo
+from .todo import registry
+
+todo_router = APIRouter(prefix="/todos")
+
+
+@todo_router.get("/")
+async def read_todos():
+    return "hello, world"
+
+
+@todo_router.post("/")
+async def _(command: CreateTodo, anywise: FastWise):
+    print(f"at router {anywise}")
+    return await anywise.send(command)
 
 
 class AppState(ty.TypedDict):
@@ -13,7 +30,13 @@ class AppState(ty.TypedDict):
 
 async def lifespan(app: FastAPI) -> ty.AsyncGenerator[AppState, None]:
     anywise = Anywise()
-    # anywise.include()
+    anywise.include([registry])
+    async with anywise._dg.scope("app") as app_scope:
+        app_scope.register_dependent(anywise, Anywise)
+        engine = await app_scope.resolve(AsyncEngine)
+        print(f"id {anywise}")
+
+    await create_tables(engine)
     yield {"anywise": anywise}
 
 
@@ -21,10 +44,10 @@ def app_factory():
     VERSION = "1"
     root_path = f"/api/v{VERSION}"
     app = FastAPI(lifespan=lifespan, version=VERSION, root_path=root_path)
-
     app.include_router(todo_router)
     return app
 
 
 if __name__ == "__main__":
-    uvicorn.run(app_factory)
+    app_str = "anywise.demo.app:app_factory"
+    uvicorn.run(app_str, reload=True)