+++ anywise/demo/model.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class TodoCommand: ...
+
+
+@dataclass
+class CreateTodo(TodoCommand):
+    title: str
+    content: str